+++ src/components/AppBox.js
@@ -31,7 +31,7 @@ import Destinations from "./pages/Destinations";
 import EventTags from "./pages/EventTags";
 import UserLogs from "./pages/UserLogs";
 import PrivateTab from "./authentication/PrivateTab";
-import UserSelfInspect from "./pages/UserSelfInspect";
+import UserAccount from "./pages/UserAccount";
 
 const AppBox = () => {
 
@@ -162,7 +162,7 @@ const AppBox = () => {
         {/*Current user account info*/}
 
         <PrivateRoute exact path={urlPrefix("/my-account")} roles={["admin", "developer", "marketer"]}>
-            <UserSelfInspect />
+            <UserAccount />
         </PrivateRoute>
 
         {/*Other*/}
